quick fix

+++ tests/il/ac/shenkar/costManager/DBModelTest.java
@@ -1,6 +1,7 @@
 package il.ac.shenkar.costManager;
 
 import org.junit.jupiter.api.*;
+import org.junit.jupiter.api.function.Executable;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
@@ -33,7 +34,8 @@ public class DBModelTest {
     @Test
     @Order(1)
     void testAddUser() throws CostManagerException {
-        User user = model.register("test_first_name", "test_last_name", "test_email", "test_password");
+        user = model.register("test_first_name", "test_last_name", "test_email", "test_password");
+        assert (user != null);
         assert (user.getUserId() != 0);
         assert (user.getName().equals("test_first_name test_last_name"));
 
@@ -49,17 +51,13 @@ public class DBModelTest {
     @Test
     @Order(3)
     void testFailedLoginWrongPassword() {
-        CostManagerException thrown = assertThrows(CostManagerException.class, () ->{
-            model.login("test_email", "wrong_password");
-        });
+        CostManagerException thrown = assertThrows(CostManagerException.class, () -> model.login("test_email", "wrong_password"));
         Assertions.assertEquals("Wrong password", thrown.getMessage());
     }
     @Test
     @Order(2)
     void testFailedLoginWrongEmail() {
-        CostManagerException thrown = assertThrows(CostManagerException.class, () ->{
-            model.login("wrong_email", "test_password");
-        });
+        CostManagerException thrown = assertThrows(CostManagerException.class, () -> model.login("wrong_email", "test_password"));
         Assertions.assertEquals("User not found", thrown.getMessage());
     }
     @Test
@@ -73,9 +71,7 @@ public class DBModelTest {
     @Test
     @Order(6)
     void createCategoryWithSameName() {
-        CostManagerException thrown = assertThrows(CostManagerException.class, () ->{
-            model.createCategory("test_category", user);
-        });
+        CostManagerException thrown = assertThrows(CostManagerException.class, () -> model.createCategory("test_category", user));
         Assertions.assertEquals("Category already exists"   , thrown.getMessage());
     }
     @Test
@@ -86,14 +82,15 @@ public class DBModelTest {
         try {
             user2 = model.register("test_first_name2", "test_last_name2", "test_email2", "test_password2");
             User user3 = user2;
-            Assertions.assertDoesNotThrow(() -> {
-                category2.set(model.createCategory("test_category", user3));
-            });
+            Assertions.assertDoesNotThrow(() -> category2.set(model.createCategory("test_category", user3)));
         } finally {
             if (user2 != null) {
                 Connection connection = DriverManager.getConnection(connectionString, USER_NAME, PASSWORD);
                 if(category2.get() != null) {
                     connection.createStatement().execute("DELETE FROM categories WHERE category_id = " + category2.get().getId());
+                }
+                for(Category c : model.getCategories(user2)) {
+                    connection.createStatement().execute("DELETE FROM categories WHERE category_id = " + c.getId());
                 }
                 connection.createStatement().execute("DELETE FROM users WHERE user_id = " + user2.getUserId());
                 connection.close();
@@ -118,11 +115,20 @@ public class DBModelTest {
         assert (items.get(0).getCategory().getId() == category.getId());
     }
     @AfterAll
-    public static void close() throws SQLException {
+    public static void close() throws SQLException, CostManagerException {
         // sql query to delete user with email test_email
         Connection connection = DriverManager.getConnection(connectionString, USER_NAME, PASSWORD);
-        connection.createStatement().executeUpdate("DELETE FROM items WHERE id = " + item.getId());
-        connection.createStatement().executeUpdate("DELETE FROM categories WHERE category_name = 'test_category'");
+        if(item != null) {
+            connection.createStatement().executeUpdate("DELETE FROM items WHERE id = " + item.getId());
+        }
+        if(user != null) {
+            for (Category c : model.getCategories(user)) {
+                connection.createStatement().execute("DELETE FROM categories WHERE category_id = " + c.getId());
+            }
+        }
+        if(category != null) {
+            connection.createStatement().executeUpdate("DELETE FROM categories WHERE category_name = 'test_category'");
+        }
         connection.createStatement().executeUpdate("DELETE FROM users WHERE email = 'test_email'");
         connection.close();
     }
